checkbox value reflection in submission.
Consider a form like this:

```javascript
<form is='xp-form' content-type='application/json'>
	<mat-checkbox name='roles[]' value='adminstrator' checked></mat-checkbox>
	<mat-checkbox name='roles[]' value='user'></mat-checkbox>
	<mat-checkbox name='roles[]' value='guest' checked></mat-checkbox>
</form>
```

Upon submission, roles was filled in with
roles = [on, on] (or [true, true] for casted form)

This makes no sense, because a checkbox is supposed to submit the 'value'
attribute not the toggle state.  This was happening because the
value attribute was not reflected into the adaptee.  This patch provides
the fix.